+++ smoke_test/steps/visitor_booking_confirmation.rb
@@ -3,7 +3,7 @@ module SmokeTest
     class VisitorBookingConfirmation < BaseStep
 
       def validate!
-        fail 'Could not find visitor booking confirmation email' if email.nil?
+        fail 'Could not find visitor booking confirmation email' unless email
       end
 
       def complete_step
@@ -15,7 +15,7 @@ module SmokeTest
       def email
         @email ||=
           with_retries do
-            mail_lookup.last_email_matching expected_email_subject
+            MailBox.find_email state.unique_email_address, expected_email_subject
           end
       end
